Lesson 18 (#17)

+++ app/controllers/comments_controller.rb
@@ -1,11 +1,12 @@
 class CommentsController < ApplicationController
-  include Questionsnswers
-  before_action :set_commentabele!
+  include QuestionsAnswers
+  before_action :set_commentable!
   before_action :set_question
+  after_action :verify_authorized
 
   def create
     @comment = @commentable.comments.build(comment_params)
-
+    authorize @comment
     if @comment.save
       flash[:success] = t '.success'
       redirect_to question_path(@question)
@@ -18,6 +19,7 @@ class CommentsController < ApplicationController
 
   def destroy
     comment = @commentable.comments.find params[:id]
+    authorize comment
     comment.destroy
     flash[:success] = t '.success'
     redirect_to question_path(@question)
